django_project/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-imagekit==4.1.0
Pillow==9.0.1
django-split-settings==1.1.0
